add options [cancelButtonText] [cancelMode] [cancelUpdateAndStopWorker] for checkVersion

+++ index.js
@@ -1,5 +1,5 @@
 import { versionTipDialog } from './components/versionTipDialog';
-import { createWorker, createWorkerFunc } from './utils/index';
+import { createWorker, createWorkerFunc, cancelUpdateFunc } from './utils/index';
 /**
  * Polling monitoring version update (No longer maintain)
  *
@@ -40,6 +40,9 @@ export const pollingCompareVersion = (localPackageVersion, originVersionFileUrl,
  * @param {string} [options.title = 'Update'] popup title (Optional)
  * @param {string} [options.description = 'V xxx is available'] popup description (Optional)
  * @param {string} [options.buttonText = 'Refresh'] popup button text (Optional)
+ * @param {string} [options.cancelButtonText] close popup button text (Optional)
+ * @param {string} [options.cancelMode = 'ignore-current-version'] close popup button mode
+ * @param {string} [options.cancelUpdateAndStopWorker = false] close popup and stop worker
  * @param {string} options.imageUrl custom popup image address (Optional)
  * @param {string} options.rocketColor custom popup rocket color in the picture (Optional)
  * @param {string} options.primaryColor custom popup primary color, act on image background color and button background color (Optional)
@@ -55,21 +58,31 @@ export const checkVersion = (config, options) => {
     worker.postMessage({
         'version-key': config.localPackageVersion,
         'polling-time': config.pollingTime || 5000,
-        'immediate': config.immediate || false,
+        immediate: config.immediate || false,
         'origin-version-file-url': config.originVersionFileUrl,
     });
     worker.onmessage = (event) => {
+        var _a;
+        // cancelMode
+        const cancelUpdateLock = cancelUpdateFunc(options === null || options === void 0 ? void 0 : options.cancelMode, (_a = event.data) === null || _a === void 0 ? void 0 : _a.refreshPageVersion, options === null || options === void 0 ? void 0 : options.cancelUpdateAndStopWorker, worker);
+        if (cancelUpdateLock)
+            return;
+        localStorage.removeItem('version-rocket:cancelled');
+        sessionStorage.removeItem('version-rocket:cancelled');
+ 
         // custom version tip UI
         if (typeof config.onVersionUpdate === 'function') {
             config.onVersionUpdate(event.data);
         }
         else {
             // default version tip ui
-            const { title, description, buttonText, imageUrl, rocketColor, primaryColor, buttonStyle, } = options || {};
+            const { title, description, buttonText, cancelButtonText, cancelMode, imageUrl, rocketColor, primaryColor, buttonStyle, } = options || {};
             versionTipDialog({
                 title,
                 description,
                 buttonText,
+                cancelButtonText,
+                cancelMode,
                 imageUrl,
                 rocketColor,
                 primaryColor,
@@ -82,8 +95,10 @@ export const checkVersion = (config, options) => {
 /**
  * destroy checkVersion
  */
-export const unCheckVersion = ({ closeDialog = false }) => {
-    worker === null || worker === void 0 ? void 0 : worker.terminate();
+export const unCheckVersion = ({ closeDialog = false, closeWorker = true }) => {
+    if (closeWorker) {
+        worker === null || worker === void 0 ? void 0 : worker.terminate();
+    }
     if (closeDialog) {
         const dialogElement = document.querySelector('#version-rocket');
         const dialogElementParent = dialogElement === null || dialogElement === void 0 ? void 0 : dialogElement.parentElement;
